extract dev tools editors into components/devtools
- LlamaPoolEditor.jsx and BuildingsEditor.jsx: presentational modals;
  state stays lifted in App because editableBuildings is also pushed
  by Save-as-Defaults
- TechTreeEditor.jsx: owns the SVG connector measurement (ref, state,
  ResizeObserver effect move out of App)
- App.jsx: 2,419 -> 1,942 lines, behaviour preserved

Closes #28
